backend/requirements.txt
asgiref
boto3
botocore
certifi
cffi
charset-normalizer
cryptography
dj-database-url
Django
django-cors-headers
django-environ
django-filter
django-rest-framework
django-storages
djangorestframework
djangorestframework-simplejwt
idna
jmespath
psycopg2-binary
PyJWT
python-dateutil
razorpay
requests
s3transfer
sqlparse
typing-extensions
uritemplate
urllib3
whitenoise
gunicorn
Pillow
